backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
qdrant-client==1.12.1
openai==1.54.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.2
sentence-transformers==3.1.1
torch==2.5.0
transformers==4.45.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
alembic==1.13.3